Split out testing to cover service loaded plugin
