+++ officefloor/core/officecompiler/src/test/java/net/officefloor/compile/impl/managedfunction/method/MethodManagedFunctionBuilderTest.java
@@ -17,18 +17,11 @@
  */
 package net.officefloor.compile.impl.managedfunction.method;
 
-import net.officefloor.compile.spi.managedfunction.source.ManagedFunctionFlowTypeBuilder;
-import net.officefloor.compile.test.managedfunction.MockAsynchronousFlow;
+import java.lang.reflect.Method;
+
 import net.officefloor.compile.test.managedfunction.clazz.MethodManagedFunctionBuilderUtil;
 import net.officefloor.compile.test.managedfunction.clazz.MethodManagedFunctionBuilderUtil.MethodResult;
-import net.officefloor.frame.api.build.Indexed;
-import net.officefloor.frame.api.function.AsynchronousFlow;
-import net.officefloor.frame.api.function.FlowCallback;
-import net.officefloor.frame.api.function.ManagedFunctionContext;
-import net.officefloor.frame.internal.structure.Flow;
-import net.officefloor.frame.test.Closure;
 import net.officefloor.frame.test.OfficeFrameTestCase;
-import net.officefloor.plugin.clazz.FlowInterface;
 import net.officefloor.plugin.managedfunction.method.MethodManagedFunctionBuilder;
 
 /**
@@ -39,7 +32,7 @@ import net.officefloor.plugin.managedfunction.method.MethodManagedFunctionBuilde
 public class MethodManagedFunctionBuilderTest extends OfficeFrameTestCase {
 
 	/**
-	 * Ensure no parameters and no return.
+	 * Ensure can invoke {@link Method} with no parameters and no return.
 	 */
 	public void testNoParametersAndNoReturn() {
 		NoParametersAndNoReturnFunction instance = new NoParametersAndNoReturnFunction();
@@ -59,7 +52,25 @@ public class MethodManagedFunctionBuilderTest extends OfficeFrameTestCase {
 	}
 
 	/**
-	 * Ensure static.
+	 * Ensure can obtain default return value.
+	 */
+	public void testReturnValue() {
+		MethodResult result = MethodManagedFunctionBuilderUtil.runMethod(new ReturnValueFunction(), "returnValue",
+				(type) -> type.setReturnType(String.class), null);
+		assertEquals("Incorrect return value", ReturnValueFunction.RETURN_VALUE, result.getReturnValue());
+	}
+
+	public static class ReturnValueFunction {
+
+		public static final String RETURN_VALUE = "RETURN";
+
+		public String returnValue() {
+			return RETURN_VALUE;
+		}
+	}
+
+	/**
+	 * Ensure invoke static {@link Method}.
 	 */
 	public void testStatic() {
 		StaticFunction.isRun = false; // ensure correct state
@@ -79,7 +90,7 @@ public class MethodManagedFunctionBuilderTest extends OfficeFrameTestCase {
 	}
 
 	/**
-	 * Ensure dependency.
+	 * Ensure inject dependency.
 	 */
 	public void testDependency() {
 		DependencyFunction dependency = new DependencyFunction();
@@ -99,144 +110,6 @@ public class MethodManagedFunctionBuilderTest extends OfficeFrameTestCase {
 		public void method(String dependency) {
 			this.object = dependency;
 		}
-	}
-
-	/**
-	 * Ensure able to inject {@link ManagedFunctionContext}.
-	 */
-	public void testManagedFunctionContext() {
-		Closure<ManagedFunctionContext<?, ?>> managedFunctionContext = new Closure<>();
-		ManagedFunctionContextFunction instance = new ManagedFunctionContextFunction();
-		MethodManagedFunctionBuilderUtil.runMethod(instance, "method", null, (context) -> {
-			managedFunctionContext.value = (ManagedFunctionContext<?, ?>) context;
-		});
-		assertSame("Should inject the managed function context", managedFunctionContext.value, instance.context);
-	}
-
-	public static class ManagedFunctionContextFunction {
-
-		public ManagedFunctionContext<?, ?> context;
-
-		public void method(ManagedFunctionContext<?, ?> context) {
-			this.context = context;
-		}
-	}
-
-	public void testAsynchronousFlow() {
-		AsynchronousFlowFunction instance = new AsynchronousFlowFunction();
-		MethodResult result = MethodManagedFunctionBuilderUtil.runMethod(instance, "method", null, null);
-		MockAsynchronousFlow[] asyncFlows = result.getAsynchronousFlows();
-		assertEquals("Should have async flow", 1, asyncFlows.length);
-		assertSame("Incorrect async flow", asyncFlows[0], instance.asyncFlow);
-	}
-
-	public static class AsynchronousFlowFunction {
-
-		public AsynchronousFlow asyncFlow;
-
-		public void method(AsynchronousFlow asyncFlow) {
-			this.asyncFlow = asyncFlow;
-		}
-	}
-
-	/**
-	 * Ensure {@link FlowInterface} with single {@link Flow}.
-	 */
-	public void testSingleFlowInterface() {
-		Closure<Object> parameter = new Closure<>();
-		Closure<FlowCallback> callback = new Closure<>();
-		SingleFlowInterfaceFunction instance = new SingleFlowInterfaceFunction();
-		MethodManagedFunctionBuilderUtil.runMethod(instance, "method", (type) -> {
-			ManagedFunctionFlowTypeBuilder<Indexed> flow = type.addFlow();
-			flow.setArgumentType(String.class);
-			flow.setLabel("single");
-		}, (context) -> {
-			context.setFlow(0, (flowParameter, flowCallback) -> {
-				parameter.value = flowParameter;
-				callback.value = flowCallback;
-			});
-		});
-		assertSame("Incorrect parameter", SingleFlowInterfaceFunction.PARAMETER, parameter.value);
-		assertSame("Incorrect callback", instance, callback.value);
-	}
-
-	public static class SingleFlowInterfaceFunction implements FlowCallback {
-
-		public static final String PARAMETER = "FLOW";
-
-		public void method(SingleFlow flow) {
-			flow.single(PARAMETER, this);
-		}
-
-		@Override
-		public void run(Throwable escalation) throws Throwable {
-			fail("Should not be invoked");
-		}
-	}
-
-	@FlowInterface
-	public static interface SingleFlow {
-		void single(String parameter, FlowCallback callback);
-	}
-
-	/**
-	 * Ensure {@link FlowInterface} with multiple {@link Flow}.
-	 */
-	public void testMultipleFlowInterface() {
-		MultipleFlowInterfaceFunction instance = new MultipleFlowInterfaceFunction();
-		final String[] orderedFlowNames = new String[] { "four", "one", "three", "two" };
-		final Object[] expectedParameters = new Object[] { null, "one", "three", Integer.valueOf(2) };
-		final FlowCallback[] expectedCallbacks = new FlowCallback[] { null, instance, null, null };
-		final boolean[] isInvoked = new boolean[] { false, false, false, false };
-		MethodManagedFunctionBuilderUtil.runMethod(instance, "method", (type) -> {
-			for (int i = 0; i < orderedFlowNames.length; i++) {
-				ManagedFunctionFlowTypeBuilder<Indexed> flow = type.addFlow();
-				flow.setLabel(orderedFlowNames[i]);
-				Object parameter = expectedParameters[i];
-				if (parameter != null) {
-					flow.setArgumentType(parameter.getClass());
-				}
-			}
-		}, (context) -> {
-			for (int i = 0; i < orderedFlowNames.length; i++) {
-				int flowIndex = i;
-				context.setFlow(flowIndex, (flowParameter, flowCallback) -> {
-					assertEquals("Incorrect parameter for flow " + flowIndex, expectedParameters[flowIndex],
-							flowParameter);
-					assertSame("Incorrect callback for flow " + flowIndex, expectedCallbacks[flowIndex], flowCallback);
-					isInvoked[flowIndex] = true;
-				});
-			}
-		});
-		for (int i = 0; i < isInvoked.length; i++) {
-			assertTrue("Flow " + i + " not invoked", isInvoked[i]);
-		}
-	}
-
-	public static class MultipleFlowInterfaceFunction implements FlowCallback {
-
-		public void method(MultipleFlow flow) {
-			flow.one("one", this);
-			flow.two(2, null);
-			flow.three("three");
-			flow.four();
-		}
-
-		@Override
-		public void run(Throwable escalation) throws Throwable {
-			fail("Should not be invoked");
-		}
-	}
-
-	@FlowInterface
-	public static interface MultipleFlow {
-		void one(String parameter, FlowCallback callback);
-
-		void two(Integer parameter, FlowCallback callback);
-
-		void three(String parameter);
-
-		void four();
 	}
 
 }
